perf: stabilize and optimize high-precision linear algebra benchmarks
- Implement dynamic size throttling (32x32/16x16) for HP providers in systematic benchmarks.
- Force double-precision matching during HP benchmarks using runThrottled context.
- Add transparent reporting notes for throttled benchmarks.
- Resolve critical NativeMPFR runtime failures (ArrayStoreException, ClassCastException).
- Add high-precision compliance report and verification tests.

+++ episteme-native/src/main/java/org/episteme/nativ/mathematics/numbers/real/NativeRealBig.java
@@ -41,7 +41,7 @@ public final class NativeRealBig extends Real {
         this.arena = Arena.ofAuto();
         this.precision = precision;
         this.ptr = arena.allocate(32); // Approximate size for mpfr_t
-        NativeSafe.invoke(MPFR_INIT2, ptr, precision);
+        NativeSafe.invoke(MPFR_INIT2, ptr, (int) precision);
         
         String mpfrValue = value;
         if (value.equalsIgnoreCase("nan")) mpfrValue = "@NaN@";
@@ -59,7 +59,7 @@ public final class NativeRealBig extends Real {
         this.arena = Arena.ofAuto();
         this.precision = precision;
         this.ptr = arena.allocate(32);
-        NativeSafe.invoke(MPFR_INIT2, ptr, precision);
+        NativeSafe.invoke(MPFR_INIT2, ptr, (int) precision);
     }
 
     public static NativeRealBig of(String value) {
